[Refactor] 폴더링 수정(#9)
