+++ planfit/src/main/java/com/planfit/server/common/exception/BadRequestException.java
@@ -1,8 +1,8 @@
-package com.planfit.server.exception;
+package com.planfit.server.common.exception;
 
 import com.planfit.server.common.message.ErrorMessage;
 
-public class BadRequestException extends BusinessException {
+public class BadRequestException extends PlanfitException {
 
     public BadRequestException(ErrorMessage errorMessage) {
         super(errorMessage);
