+++ planfit/src/main/java/com/planfit/server/common/exception/PlanfitException.java
@@ -1,14 +1,14 @@
-package com.planfit.server.exception;
+package com.planfit.server.common.exception;
 
 import com.planfit.server.common.message.ErrorMessage;
 import lombok.Getter;
 
 @Getter
-public class BusinessException extends RuntimeException {
+public class PlanfitException extends RuntimeException {
 
     private ErrorMessage errorMessage;
 
-    public BusinessException(ErrorMessage errorMessage) {
+    public PlanfitException(ErrorMessage errorMessage) {
         super(errorMessage.getMessage());
         this.errorMessage = errorMessage;
     }
