+++ planfit/src/main/java/com/planfit/server/common/handler/GlobalExceptionHandler.java
@@ -1,7 +1,8 @@
-package com.planfit.server.common;
+package com.planfit.server.common.handler;
 
-import com.planfit.server.exception.BadRequestException;
-import lombok.RequiredArgsConstructor;
+import com.planfit.server.common.ApiResponseUtil;
+import com.planfit.server.common.BaseResponse;
+import com.planfit.server.common.exception.BadRequestException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
